pyproject.toml
[project]
name = "efficient-micro-net"
version = "1.0.0"
description = "EfficientMicroNet: A Novel Lightweight Image Classification Architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Research Implementation"}
]
keywords = ["deep-learning", "image-classification", "lightweight", "efficient", "pytorch"]

dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "thop>=0.1.1",
    "tensorboard>=2.11.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "utils*"]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
